Show switch offer number. Split BerthSwitchOfferCard to container and presentational components
