+++ src/locales/fi.json
@@ -126,6 +126,7 @@
       "harbor": "Satama",
       "harborChief": "Satamapäällikkö",
       "harborMap": "Satamakartta",
+      "id": "Tunniste",
       "inQueue": "Jonossa",
       "invoice": "Lasku",
       "invoices": "Laskut",
